Refactored extended BMP into distinct module
