+++ Texconv/PortablePixMap.cpp
@@ -82,8 +82,12 @@ namespace
     {
         blob.reset();
 
+#if (_WIN32_WINNT >= _WIN32_WINNT_WIN8)
+        ScopedHandle hFile(safe_handle(CreateFile2(szFile, GENERIC_READ, FILE_SHARE_READ, OPEN_EXISTING, nullptr)));
+#else
         ScopedHandle hFile(safe_handle(CreateFileW(szFile, GENERIC_READ, FILE_SHARE_READ, nullptr, OPEN_EXISTING,
             FILE_FLAG_SEQUENTIAL_SCAN, nullptr)));
+#endif
         if (!hFile)
         {
             return HRESULT_FROM_WIN32(GetLastError());
@@ -374,7 +378,11 @@ HRESULT __cdecl SaveToPortablePixMap(
         }
     }
 
+#if (_WIN32_WINNT >= _WIN32_WINNT_WIN8)
+    ScopedHandle hFile(safe_handle(CreateFile2(szFile, GENERIC_WRITE, 0, CREATE_ALWAYS, nullptr)));
+#else
     ScopedHandle hFile(safe_handle(CreateFileW(szFile, GENERIC_WRITE, 0, nullptr, CREATE_ALWAYS, 0, nullptr)));
+#endif
     if (!hFile)
         return HRESULT_FROM_WIN32(GetLastError());
 
@@ -563,7 +571,11 @@ HRESULT __cdecl SaveToPortablePixMapHDR(
 
     tmpImage.Release();
 
+#if (_WIN32_WINNT >= _WIN32_WINNT_WIN8)
+    ScopedHandle hFile(safe_handle(CreateFile2(szFile, GENERIC_WRITE, 0, CREATE_ALWAYS, nullptr)));
+#else
     ScopedHandle hFile(safe_handle(CreateFileW(szFile, GENERIC_WRITE, 0, nullptr, CREATE_ALWAYS, 0, nullptr)));
+#endif
     if (!hFile)
         return HRESULT_FROM_WIN32(GetLastError());
